Separate class definition and method implementation
